+++ src/sqlite/repository.rs
@@ -608,7 +608,7 @@ impl crate::repository::LnurlRepository for LnurlRepository {
             account_id
         };
 
-        sqlx::query(
+        let update_result = sqlx::query(
             "UPDATE account_identifiers
              SET account_id = $3
              ,   description = $4
@@ -623,6 +623,9 @@ impl crate::repository::LnurlRepository for LnurlRepository {
         .bind(&transfer.source_account_id)
         .execute(&mut *tx)
         .await?;
+        if update_result.rows_affected() != 1 {
+            return Err(LnurlRepositoryError::SourceNotOwner);
+        }
 
         sqlx::query(
             "DELETE FROM account_identifiers
